+++ test/support/fake_pipedrive_server.ex
@@ -23,7 +23,9 @@ defmodule ExPipedrive.FakePipedriveServer do
   import ExPipedrive.FakeFollowerV2ApiHandler
   import ExPipedrive.FakeGoalApiHandler
   import ExPipedrive.FakeLeadApiHandler
+  import ExPipedrive.FakeLeadFieldApiHandler
   import ExPipedrive.FakeLeadLabelApiHandler
+  import ExPipedrive.FakeLeadSourceApiHandler
   import ExPipedrive.FakeMailboxApiHandler
   import ExPipedrive.FakeNoteApiHandler
   import ExPipedrive.FakeNoteFieldApiHandler
@@ -241,6 +243,18 @@ defmodule ExPipedrive.FakePipedriveServer do
     conn
     |> put_resp_header("content-type", "application/json;charset=utf-8")
     |> handle_list_pipelines()
+  end
+
+  get "/api/v1/leadFields" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_list_lead_fields(conn.query_params)
+  end
+
+  get "/api/v1/leadSources" do
+    conn
+    |> put_resp_header("content-type", "application/json;charset=utf-8")
+    |> handle_list_lead_sources()
   end
 
   get "/api/v1/leadLabels" do